% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_request.R
\name{bad_getChQuery}
\alias{bad_getChQuery}
\title{Execute SQL request at remote ClickHouse, analyze and process errors in Shiny context}
\usage{
bad_getChQuery(conn, req)
}
\arguments{
\item{conn}{ClickHouse connection object}

\item{req}{Formatted SQL request (string)}
}
\value{
Data.frame recieved from ClickHouse
}
\description{
All
}
