% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_request.R
\name{build_request}
\alias{build_request}
\alias{buildReqFilter}
\alias{buildReqLimits}
\title{Set of helpers to build SQL request}
\usage{
buildReqFilter(field, conds, add = TRUE)

buildReqLimits(begin, end, region = NULL, prefix = NULL, channel = NULL,
  event = NULL, segment = NULL, serial_mask = NULL)
}
\arguments{
\item{field}{DB field to include in filter restriction}

\item{conds}{Character vector of permitted field values. In case conditions
are absent (NULL/NA or 'all' value) any restrictions will be omitted}

\item{add}{Type of unfolding. TRUE value shows that filter will be a part of
complex restriction and should be ADDED}

\item{begin}{Start date-range}

\item{end}{End of date-range}

\item{...}{Custom named params}
}
\value{
Unfolded into string filter condition

Unfolded into string filter condition
}
\description{
Set of helpers to build SQL request

Build separate subcondition of SQL request filter
Conditions can be vector.

Build SQL request restriction based on user-defined set of fields
}
