% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_request.R
\name{buildReqFilter}
\alias{buildReqFilter}
\title{Construct separate condition of SQL request filter}
\usage{
buildReqFilter(field, conds, add = TRUE)
}
\arguments{
\item{field}{DB field to include in filter restriction}

\item{conds}{Character vector of permitted field values. In case conditions
are absent (NULL/NA or 'all' value) any restrictions will be omitted}

\item{add}{Type of unfolding. TRUE value shows that filter will be a part of
complex restriction and should be added}
}
\value{
Filter condition unfolded into string
}
\description{
Construct separate condition of SQL request filter.
Conditions can be vector.
}
