% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_request.R
\name{buildReqLimits}
\alias{buildReqLimits}
\title{Build SQL request restriction based on user-defined set of fields}
\usage{
buildReqLimits(begin, end, region = NULL, prefix = NULL, channel = NULL,
  event = NULL, segment = NULL, serial_mask = "")
}
\arguments{
\item{begin}{Start date (strictly Date class)}

\item{end}{End date, (strictly Date class)}

\item{region}{Region filter (string vector)}

\item{prefix}{Prefix filter (string vector)}

\item{channel}{Channel filter (string vector)}

\item{event}{Event filter (string vector)}

\item{segment}{Segment filter (string vector)}

\item{serial_mask}{String to find in serial number whithin \%LIKE\% condition}
}
\value{
Limits unfolded into string
}
\description{
All
}
