% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_group.R
\name{hgroup.enum}
\alias{hgroup.enum}
\title{Arrange time vector according equidistant time intervals}
\usage{
hgroup.enum(date, hours_bin = NULL, mins_bin = 5)
}
\arguments{
\item{date}{POSIXct vector to arrange}

\item{hours_bin}{Duration (in hours) between arrangement points}

\item{mins_bin}{Duration (in minutes) between arrangement points}
}
\description{
There are two possible ways of arrangement: by hour's intervals OR by
minute's interval. Hour's arrangement has high priority. Hour arrangement
must be integer (1, 2, ...) or fractional part of hour (0.25 or 0.5 only)
}
