% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_styles.R
\name{initRCfontInCairo}
\alias{initRCfontInCairo}
\title{Cairo initialization in linux with Roboto Condensed font}
\usage{
initRCfontInCairo()
}
\description{
Cairo initialization in linux with Roboto Condensed font
}
