% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_styles.R
\name{theme_dvt}
\alias{theme_dvt}
\title{Custom ggplot theme based on ipsum-theme}
\usage{
theme_dvt(target = "screen")
}
\arguments{
\item{target}{Target for plot. May be on of "screen" or "word_A4". Target will
affect on size, font, colour, resolution, etc.}
}
\description{
Custom ggplot theme based on ipsum-theme
}
